Fix tests
Fix docker-compose structure to run tests in different container than app to avoid data collision. Optimized structure of project to avoid duplications or unnecessary references. Tests are now working and they are not coliding with the application.

Improved the docker-compose structure to run tests in a different container than the application to avoid data collisions. Optimized the project structure to avoid duplication or unnecessary references. Tests now work and do not collide with the application.

Resolves: #12